[refactoring] 사용하지 않는 import 및 파일 삭제
