+++ app/src/main/java/kr/co/nexters/winepick/data/model/WineFood.kt
@@ -1,6 +1,6 @@
 package kr.co.nexters.winepick.data.model
 
 data class WineFood(
-    var title : String,
+    var title: String,
     var img: Int? = null
 )
