+++ app/src/main/java/kr/co/nexters/winepick/data/model/remote/wine/WineResult.kt
@@ -93,19 +93,19 @@ data class WineResult(
 
     @SerializedName("likeYn")
     @SerialName("likeYn")
-    var likeYn : Boolean? = null,
+    var likeYn: Boolean? = null,
 
     @SerializedName("purpose")
     @SerialName("purpose")
-    val purpose : String? = null,
+    val purpose: String? = null,
 
     @SerializedName("store")
     @SerialName("store")
-    val store : String? = null,
+    val store: String? = null,
 
-    var wineImg : Int? = null,
+    var wineImg: Int? = null,
 
-    var wineCountryImg : Int? = null
+    var wineCountryImg: Int? = null
 ) : java.io.Serializable {
     /** 만약 해당 [WineResult] 아이템에 좋아요 버튼 클릭 시 새로 설정될 [likeYn] */
     val clickedLikeYn: Boolean
