+++ app/src/main/java/kr/co/nexters/winepick/data/model/remote/wine/WineValue.kt
@@ -1,10 +1,10 @@
 package kr.co.nexters.winepick.data.model.remote.wine
 
 data class WineValue(
-        val wineTitle: String,
-        val wineDetail : String,
-        val wineNumber : Int,
-        val wineHigh : String,
-        val wineLow : String,
-        var hintVisible : Boolean = false
+    val wineTitle: String,
+    val wineDetail: String,
+    val wineNumber: Int,
+    val wineHigh: String,
+    val wineLow: String,
+    var hintVisible: Boolean = false
 )
