+++ app/src/main/java/kr/co/nexters/winepick/data/model/remote/wine/WinesResult.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
-import kr.co.nexters.winepick.data.constant.DependencyServer
+import kr.co.nexters.winepick.constant.DependencyServer
 import kr.co.nexters.winepick.data.model.WinePickResponse
 import kr.co.nexters.winepick.network.WinePickService
