+++ app/src/main/java/kr/co/nexters/winepick/data/response/UserData.kt
@@ -1,9 +1,6 @@
 package kr.co.nexters.winepick.data.response
 
 import com.google.gson.annotations.SerializedName
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import java.util.*
 
 
 data class UserData(
