+++ app/src/main/java/kr/co/nexters/winepick/ui/detail/WineDetailViewModel.kt
@@ -1,12 +1,10 @@
 package kr.co.nexters.winepick.ui.detail
 
-import android.content.Intent
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kr.co.nexters.winepick.R
 import kr.co.nexters.winepick.WinePickApplication
-import kr.co.nexters.winepick.data.model.AccessTokenData
 import kr.co.nexters.winepick.data.model.LikeWine
 import kr.co.nexters.winepick.data.model.WineFood
 import kr.co.nexters.winepick.data.model.remote.wine.WineResult
@@ -14,9 +12,7 @@ import kr.co.nexters.winepick.data.model.remote.wine.WineValue
 import kr.co.nexters.winepick.data.repository.WinePickRepository
 import kr.co.nexters.winepick.di.AuthManager
 import kr.co.nexters.winepick.ui.base.BaseViewModel
-import kr.co.nexters.winepick.ui.home.HomeActivity
 import timber.log.Timber
-import java.lang.StringBuilder
 import javax.inject.Inject
 
 /**
