+++ app/src/main/java/kr/co/nexters/winepick/ui/detail/WineValueAdapter.kt
@@ -1,27 +1,20 @@
 package kr.co.nexters.winepick.ui.detail
 
-import android.os.Handler
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import android.view.animation.AnimationUtils
-import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import kr.co.nexters.winepick.R
-import kr.co.nexters.winepick.WinePickApplication
-import kr.co.nexters.winepick.data.model.WineFood
-import kr.co.nexters.winepick.data.model.remote.wine.WineResult
 import kr.co.nexters.winepick.data.model.remote.wine.WineValue
-import kr.co.nexters.winepick.databinding.ItemWineResultBinding
 import kr.co.nexters.winepick.databinding.ItemWineValueBinding
-import kr.co.nexters.winepick.databinding.ItemWineValueBindingImpl
-import kr.co.nexters.winepick.ui.base.WineResultViewModel
-import kr.co.nexters.winepick.ui.search.WineResultViewHolder
-import kr.co.nexters.winepick.util.setOnSingleClickListener
 
-
+/**
+ * WineValueAdapter
+ *
+ * @since v1.0.0 / 2021.03.30
+ */
 class WineValueAdapter(val vm: WineDetailViewModel) :
     ListAdapter<WineValue, WineValueViewHolder>(WineValueDiffUtilCallBack) {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): WineValueViewHolder {
