+++ app/src/main/java/kr/co/nexters/winepick/ui/type/TypeDetailActivity.kt
@@ -6,25 +6,20 @@ import android.view.View
 import androidx.activity.viewModels
 import androidx.lifecycle.Observer
 import com.kakao.sdk.auth.LoginClient
-import com.kakao.sdk.auth.model.OAuthToken
-import com.kakao.sdk.user.UserApiClient
 import dagger.hilt.android.AndroidEntryPoint
 import kr.co.nexters.winepick.BR
 import kr.co.nexters.winepick.R
 import kr.co.nexters.winepick.WinePickApplication
-import kr.co.nexters.winepick.data.constant.Constant
+import kr.co.nexters.winepick.constant.Constant
 import kr.co.nexters.winepick.data.model.remote.wine.WineResult
 import kr.co.nexters.winepick.data.repository.WineRepository
 import kr.co.nexters.winepick.databinding.ActivityTypeDetailBinding
-import kr.co.nexters.winepick.di.AuthManager
 import kr.co.nexters.winepick.type.RecentSearchAdapter
 import kr.co.nexters.winepick.ui.base.BaseActivity
 import kr.co.nexters.winepick.ui.component.ConfirmDialog
 import kr.co.nexters.winepick.ui.detail.WineDetailActivity
-import kr.co.nexters.winepick.ui.login.LoginViewModel
 import kr.co.nexters.winepick.ui.survey.SurveyActivity
 import kr.co.nexters.winepick.util.VerticalItemDecorator
-import timber.log.Timber
 import javax.inject.Inject
 
 @AndroidEntryPoint
