+++ app/src/main/java/kr/co/nexters/winepick/ui/webview/WebViewActivity.kt
@@ -6,7 +6,7 @@ import android.webkit.WebView
 import android.webkit.WebViewClient
 import dagger.hilt.android.AndroidEntryPoint
 import kr.co.nexters.winepick.R
-import kr.co.nexters.winepick.data.constant.Constant
+import kr.co.nexters.winepick.constant.Constant
 import kr.co.nexters.winepick.databinding.ActivityWebViewBinding
 import kr.co.nexters.winepick.ui.base.BaseActivity
 import kr.co.nexters.winepick.ui.base.BaseViewModel
