--- app/src/main/java/kr/co/nexters/winepick/util/KotlinxSerializerModules.kt
@@ -1,9 +1,0 @@
-package kr.co.nexters.winepick.util
-
-/**
- * kotlinx.serialization 을 사용하기 위해 만들어둔 모듈
- *
- * @see [kotlinx.serialization 공식 문서](https://github.com/Kotlin/kotlinx.serialization/blob/master/docs/serialization-guide.md)
- *
- * @since v1.0.0 / 2021.02.02
- */
