+++ app/src/main/java/kr/co/nexters/winepick/util/VerticalItemDecorator.kt
@@ -4,7 +4,6 @@ import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
-
 class VerticalItemDecorator(private val divHeight : Int) : RecyclerView.ItemDecoration() {
 
     override fun getItemOffsets(outRect: Rect, view: View, parent : RecyclerView, state : RecyclerView.State) {
