+++ app/src/test/java/kr/co/nexters/winepick/data/repository/SurveyRepositoryTest.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
 import kr.co.nexters.winepick.base.AndroidBaseTest
-import kr.co.nexters.winepick.data.constant.Constant
+import kr.co.nexters.winepick.constant.Constant
 import kr.co.nexters.winepick.data.model.Survey
 import kr.co.nexters.winepick.data.model.SurveyAnswerType
 import org.junit.Assert
