+++ app/src/test/java/kr/co/nexters/winepick/data/repository/WineRepositoryTest.kt
@@ -4,7 +4,7 @@ import dagger.hilt.android.testing.HiltAndroidTest
 import kotlinx.serialization.builtins.ListSerializer
 import kotlinx.serialization.json.Json
 import kr.co.nexters.winepick.base.AndroidBaseTest
-import kr.co.nexters.winepick.data.constant.Constant
+import kr.co.nexters.winepick.constant.Constant
 import kr.co.nexters.winepick.data.model.remote.wine.WineResult
 import kr.co.nexters.winepick.data.model.remote.wine.getWineResponse
 import org.junit.Assert
